requirements.txt
fastapi
uvicorn
requests
gtts
moviepy
pillow
numpy
google-generativeai